Separate FlitchioStatusListener and InternalStatusListener to avoid confusion
