+++ sdk/src/main/java/com/supenta/flitchio/sdk/InternalStatusListener.java
@@ -1,0 +1,5 @@
+package com.supenta.flitchio.sdk;
+
+interface InternalStatusListener {
+    void onInternalStatusChanged(Status status);
+}
